+++ integration/container/ps_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/docker/docker/testutil"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
+	"gotest.tools/v3/skip"
 )
 
 func TestPsFilter(t *testing.T) {
@@ -46,4 +47,30 @@ func TestPsFilter(t *testing.T) {
 		assert.NilError(t, err)
 		assert.Check(t, is.Contains(containerIDs(results), prev))
 	})
+}
+
+// TestPsPlatform verifies that containers have a platform set
+func TestPsImageManifestPlatform(t *testing.T) {
+	skip.If(t, testEnv.IsRemoteDaemon)
+	skip.If(t, testEnv.DaemonInfo.OSType != "linux")
+	skip.If(t, !testEnv.UsingSnapshotter())
+
+	ctx := setupTest(t)
+	apiClient := testEnv.APIClient()
+
+	container.Create(ctx, t, apiClient)
+
+	containers, err := apiClient.ContainerList(ctx, containertypes.ListOptions{
+		All: true,
+	})
+	assert.NilError(t, err)
+	assert.Check(t, len(containers) > 0)
+
+	ctr := containers[0]
+	if assert.Check(t, ctr.ImageManifestDescriptor != nil && ctr.ImageManifestDescriptor.Platform != nil) {
+		// Check that at least OS and Architecture have a value. Other values
+		// depend on the platform on which we're running the test.
+		assert.Equal(t, ctr.ImageManifestDescriptor.Platform.OS, testEnv.DaemonInfo.OSType)
+		assert.Check(t, ctr.ImageManifestDescriptor.Platform.Architecture != "")
+	}
 }